requirements.txt
Django==5.0.7
requests==2.23.0
PyMySQL==0.9.3
djangorestframework==3.15.2
pygments==2.9.0
snowland-smx==0.3.1
django-redis==5.4.0
redis==5.0.4
django-filter==21.1
djangorestframework-simplejwt==5.3.1